requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Data Management
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2  # Per export Excel

# Environment
python-dotenv==1.0.0

# Logging e Utility
python-json-logger==2.0.7

# Database (per future espansioni)
SQLAlchemy==2.0.23
alembic==1.13.0

# Testing (opzionale)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development (opzionale)
black==23.12.0
flake8==7.0.0
mypy==1.7.1